Rework: move chainspec generation to separate module
