+++ bin/node/src/lib.rs
@@ -7,6 +7,7 @@ mod executor;
 mod resources;
 mod rpc;
 mod service;
+mod shared_params;
 
 pub use cli::{Cli, Subcommand};
 pub use executor::ExecutorDispatch;
