+++ bin/node/src/service.rs
@@ -28,9 +28,9 @@ use sp_arithmetic::traits::BaseArithmetic;
 use sp_consensus::DisableProofRecording;
 use sp_consensus_aura::{sr25519::AuthorityPair as AuraPair, Slot};
 
+use crate::shared_params::DEFAULT_BACKUP_FOLDER;
 use crate::{
     aleph_cli::AlephCli,
-    chain_spec::DEFAULT_BACKUP_FOLDER,
     executor::AlephExecutor,
     rpc::{create_full as create_full_rpc, FullDeps as RpcFullDeps},
 };
